ref: Rearchitect the log and slog Integrations
