+++ sentry-log/src/integration.rs
@@ -1,28 +1,65 @@
+#![allow(deprecated)]
+
 use std::sync::Once;
 
-use log::{Level, LevelFilter, Record};
+use log::Record;
+use sentry_core::protocol::{Breadcrumb, Event};
 use sentry_core::{ClientOptions, Integration};
 
 use crate::logger::Logger;
+
+/// The Action that Sentry should perform for a [`log::Metadata`].
+#[derive(Debug)]
+pub enum LevelFilter {
+    /// Ignore the [`Record`].
+    Ignore,
+    /// Create a [`Breadcrumb`] from this [`Record`].
+    Breadcrumb,
+    /// Create a message [`Event`] from this [`Record`].
+    Event,
+    /// Create an exception [`Event`] from this [`Record`].
+    Exception,
+}
+
+/// The type of Data Sentry should ingest for a [`log::Record`].
+#[allow(clippy::large_enum_variant)]
+pub enum RecordMapping {
+    /// Ignore the [`Record`]
+    Ignore,
+    /// Adds the [`Breadcrumb`] to the Sentry scope.
+    Breadcrumb(Breadcrumb),
+    /// Captures the [`Event`] to Sentry.
+    Event(Event<'static>),
+}
 
 /// Logger specific options.
 pub struct LogIntegration {
     /// The global filter that should be used (also used before dispatching
     /// to the nested logger).
-    pub global_filter: Option<LevelFilter>,
+    #[deprecated = "use the [`filter()`] function instead"]
+    pub global_filter: Option<log::LevelFilter>,
     /// The sentry specific log level filter (defaults to `Info`)
-    pub filter: LevelFilter,
+    #[deprecated = "use the [`filter()`] function instead"]
+    pub filter: log::LevelFilter,
     /// If set to `true`, breadcrumbs will be emitted. (defaults to `true`)
+    #[deprecated = "use the [`filter()`] function instead"]
     pub emit_breadcrumbs: bool,
     /// If set to `true` error events will be sent for errors in the log. (defaults to `true`)
+    #[deprecated = "use the [`filter()`] function instead"]
     pub emit_error_events: bool,
     /// If set to `true` warning events will be sent for warnings in the log. (defaults to `false`)
+    #[deprecated = "use the [`filter()`] function instead"]
     pub emit_warning_events: bool,
     /// If set to `true` current stacktrace will be resolved and attached
     /// to each event. (expensive, defaults to `true`)
+    #[deprecated = "use builder functions instead; direct field access will be removed soon"]
     pub attach_stacktraces: bool,
     /// The destination log.
+    #[deprecated = "use builder functions instead; direct field access will be removed soon"]
     pub dest_log: Option<Box<dyn log::Log>>,
+
+    sentry_filter: Option<Box<dyn Fn(&log::Metadata<'_>) -> LevelFilter + Send + Sync>>,
+    mapper: Option<Box<dyn Fn(&Record<'_>) -> RecordMapping + Send + Sync>>,
 }
 
 static INIT: Once = Once::new();
@@ -52,12 +89,14 @@ impl Default for LogIntegration {
     fn default() -> Self {
         Self {
             global_filter: None,
-            filter: LevelFilter::Info,
+            filter: log::LevelFilter::Info,
             emit_breadcrumbs: true,
             emit_error_events: true,
             emit_warning_events: false,
             attach_stacktraces: true,
             dest_log: None,
+            sentry_filter: None,
+            mapper: None,
         }
     }
 }
@@ -81,6 +120,11 @@ impl std::fmt::Debug for LogIntegration {
 }
 
 impl LogIntegration {
+    /// Creates a new `log` Integration.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
     /// Initializes an env logger as destination target.
     #[cfg(feature = "env_logger")]
     pub fn with_env_logger_dest(mut self, logger: Option<env_logger::Logger>) -> Self {
@@ -94,13 +138,37 @@ impl LogIntegration {
         self
     }
 
+    /// Sets a custom filter function.
+    ///
+    /// The filter classifies how sentry should handle [`Record`]s based on
+    /// their [`log::Metadata`].
+    pub fn filter<F>(mut self, filter: F) -> Self
+    where
+        F: Fn(&log::Metadata<'_>) -> LevelFilter + Send + Sync + 'static,
+    {
+        self.sentry_filter = Some(Box::new(filter));
+        self
+    }
+
+    /// Sets a custom mapper function.
+    ///
+    /// The mapper is responsible for creating either breadcrumbs or events
+    /// from [`Record`]s.
+    pub fn mapper<M>(mut self, mapper: M) -> Self
+    where
+        M: Fn(&Record<'_>) -> RecordMapping + Send + Sync + 'static,
+    {
+        self.mapper = Some(Box::new(mapper));
+        self
+    }
+
     /// Returns the effective global filter.
     ///
     /// This is what is set for these logger options when the log level
     /// needs to be set globally.  This is the greater of `global_filter`
     /// and `filter`.
     #[inline(always)]
-    pub(crate) fn effective_global_filter(&self) -> LevelFilter {
+    pub(crate) fn effective_global_filter(&self) -> log::LevelFilter {
         let filter = if let Some(filter) = self.global_filter {
             if filter < self.filter {
                 self.filter
@@ -117,21 +185,21 @@ impl LogIntegration {
     ///
     /// This is controlled by `emit_error_events` and `emit_warning_events`.
     #[inline(always)]
-    fn issue_filter(&self) -> LevelFilter {
+    fn issue_filter(&self) -> log::LevelFilter {
         if self.emit_warning_events {
-            LevelFilter::Warn
+            log::LevelFilter::Warn
         } else if self.emit_error_events {
-            LevelFilter::Error
+            log::LevelFilter::Error
         } else {
-            LevelFilter::Off
+            log::LevelFilter::Off
         }
     }
 
     /// Checks if an issue should be created.
     pub(crate) fn create_issue_for_record(&self, record: &Record<'_>) -> bool {
         match record.level() {
-            Level::Warn => self.emit_warning_events,
-            Level::Error => self.emit_error_events,
+            log::Level::Warn => self.emit_warning_events,
+            log::Level::Error => self.emit_error_events,
             _ => false,
         }
     }
@@ -139,6 +207,8 @@ impl LogIntegration {
 
 #[test]
 fn test_filters() {
+    use log::LevelFilter;
+
     let opt_warn = LogIntegration {
         filter: LevelFilter::Warn,
         ..Default::default()
