+++ sentry-slog/src/lib.rs
@@ -1,9 +1,8 @@
 //! Sentry `slog` Integration.
 //!
-//! The sentry `slog` integration consists of two parts, the
-//! [`SlogIntegration`] which configures how sentry should treat
-//! `slog::Record`s, and the [`SentryDrain`], which can be used to create a
-//! `slog::Logger`.
+//! This mainly provides the [`SentryDrain`], which wraps another [`slog::Drain`]
+//! and can be configured to forward [`slog::Record`]s to Sentry.
+//! The [`SentryDrain`] can be used to create a `slog::Logger`.
 //!
 //! *NOTE*: This integration currently does not process any `slog::KV` pairs,
 //! but support for this will be added in the future.
@@ -14,14 +13,13 @@
 //! use sentry::{init, ClientOptions};
 //! use sentry_slog::{SentryDrain, SlogIntegration};
 //!
-//! let integration = SlogIntegration::default();
-//! let options = ClientOptions::default().add_integration(integration);
+//! let options = ClientOptions::new().add_integration(SlogIntegration::new());
 //! let _sentry = sentry::init(options);
 //!
 //! let drain = SentryDrain::new(slog::Discard);
 //! let root = slog::Logger::root(drain, slog::o!());
 //!
-//! # let options = ClientOptions::default().add_integration(SlogIntegration::default());
+//! # let options = ClientOptions::new().add_integration(SlogIntegration::new());
 //! # let events = sentry::test::with_captured_events_options(|| {
 //! slog::info!(root, "recorded as breadcrumb");
 //! slog::warn!(root, "recorded as regular event");
@@ -45,23 +43,26 @@
 //! assert_eq!(captured_event.exception.len(), 1);
 //! ```
 //!
-//! The integration can also be customized with a `filter`, and a `mapper`:
+//! The Drain can also be customized with a `filter`, and a `mapper`:
 //!
 //! ```
-//! use sentry_slog::{exception_from_record, LevelFilter, RecordMapping, SlogIntegration};
+//! use sentry_slog::{exception_from_record, LevelFilter, RecordMapping, SentryDrain};
 //!
-//! let integration = SlogIntegration::default()
+//! let drain = SentryDrain::new(slog::Discard)
 //!     .filter(|level| match level {
 //!         slog::Level::Critical | slog::Level::Error => LevelFilter::Event,
 //!         _ => LevelFilter::Ignore,
 //!     })
-//!     .mapper(|record, kv| RecordMapping::Event(exception_from_record(record, kv)));
+//!     .mapper(|record, kv| match record.level() {
+//!         slog::Level::Critical | slog::Level::Error =>
+//!             RecordMapping::Event(exception_from_record(record, kv))),
+//!         _ => RecordMapping::Ignore,
+//!     });
 //! ```
 //!
-//! Please not that the `mapper` can override any classification from the
-//! previous `filter`.
+//! When a `mapper` is specified, a corresponding `filter` should also be
+//! provided.
 //!
-//! [`SlogIntegration`]: struct.SlogIntegration.html
 //! [`SentryDrain`]: struct.SentryDrain.html
 
 #![doc(html_favicon_url = "https://sentry-brand.storage.googleapis.com/favicon.ico")]
@@ -74,5 +75,5 @@ mod drain;
 mod integration;
 
 pub use converters::*;
-pub use drain::SentryDrain;
-pub use integration::{default_filter, LevelFilter, RecordMapping, SlogIntegration};
+pub use drain::{default_filter, LevelFilter, RecordMapping, SentryDrain};
+pub use integration::SlogIntegration;
